Create Creating Zoom CSV's
We needed to create 2 zooms with pre-assigned breakout rooms every 30 minutes and only had about 10 minutes to prepare the CSV files necessary for input. This script automatically generates a sheet with the necessary format for zoom, all that's required is to download it as a CSV. The automation of this process reduced a lot of stress.